isolate code running process
